Separate Statement and Container classes
